[TRL-430] refactor: 커스텀 어노테이션 적용
