+++ src/main/java/com/cosain/trilo/trip/presentation/day/DayColorUpdateController.java
@@ -1,6 +1,8 @@
 package com.cosain.trilo.trip.presentation.day;
 
-import com.cosain.trilo.common.LoginUser;
+import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.presentation.Login;
+import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.day.dto.DayColorUpdateCommand;
 import com.cosain.trilo.trip.application.day.dto.factory.DayColorUpdateCommandFactory;
 import com.cosain.trilo.trip.application.day.service.DayColorUpdateService;
@@ -22,12 +24,13 @@ public class DayColorUpdateController {
 
     @PutMapping("/api/days/{dayId}/color")
     @ResponseStatus(HttpStatus.OK)
+    @Login
     public DayColorUpdateResponse updateDayColor(
             @PathVariable("dayId") Long dayId,
-            @LoginUser User user,
+            @LoginUser UserPayload userPayload,
             @RequestBody DayColorUpdateRequest request) {
 
-        Long tripperId = user.getId();
+        Long tripperId = userPayload.getId();
 
         DayColorUpdateCommand command = dayColorUpdateCommandFactory.createCommand(request.getColorName());
