+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripImageUpdateController.java
@@ -1,6 +1,8 @@
 package com.cosain.trilo.trip.presentation.trip;
 
-import com.cosain.trilo.common.LoginUser;
+import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.presentation.Login;
+import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.common.file.ImageFile;
 import com.cosain.trilo.trip.application.trip.service.TripImageUpdateService;
 import com.cosain.trilo.trip.presentation.trip.dto.response.TripImageUpdateResponse;
@@ -20,19 +22,20 @@ public class TripImageUpdateController {
 
     /**
      * 여행의 이미지를 변경하고, 저장된 이미지 경로를 응답합니다.
-     * @param user : 사용자
+     * @param userPayload : 토큰 페이로드
      * @param tripId : 여행 식별자(id)
      * @param multipartFile : 파일
      * @return 응답 API (여행 식별자, 이미지 경로)
      */
     @PostMapping("/api/trips/{tripId}/image/update")
     @ResponseStatus(HttpStatus.OK)
+    @Login
     public TripImageUpdateResponse updateTripImage(
-            @LoginUser User user,
+            @LoginUser UserPayload userPayload,
             @PathVariable Long tripId,
             @RequestParam("image") MultipartFile multipartFile) {
 
-        Long tripperId = user.getId();
+        Long tripperId = userPayload.getId();
         ImageFile imageFile = ImageFile.from(multipartFile);
 
         String imageURL = tripImageUpdateService.updateTripImage(tripId, tripperId, imageFile);
