+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripPeriodUpdateController.java
@@ -1,6 +1,8 @@
 package com.cosain.trilo.trip.presentation.trip;
 
-import com.cosain.trilo.common.LoginUser;
+import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.presentation.Login;
+import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.trip.service.TripPeriodUpdateService;
 import com.cosain.trilo.trip.application.trip.dto.TripPeriodUpdateCommand;
 import com.cosain.trilo.trip.application.trip.dto.factory.TripPeriodUpdateCommandFactory;
@@ -22,8 +24,9 @@ public class TripPeriodUpdateController {
 
     @PutMapping("/api/trips/{tripId}/period")
     @ResponseStatus(HttpStatus.OK)
-    public TripPeriodUpdateResponse updateTrip(@LoginUser User user, @PathVariable Long tripId, @RequestBody TripPeriodUpdateRequest request) {
-        Long tripperId = user.getId();
+    @Login
+    public TripPeriodUpdateResponse updateTrip(@LoginUser UserPayload userPayload, @PathVariable Long tripId, @RequestBody TripPeriodUpdateRequest request) {
+        Long tripperId = userPayload.getId();
         TripPeriodUpdateCommand updateCommand = tripPeriodUpdateCommandFactory.createCommand(request.getStartDate(), request.getEndDate());
 
         tripPeriodUpdateService.updateTripPeriod(tripId, tripperId, updateCommand);
