+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripTitleUpdateController.java
@@ -1,6 +1,8 @@
 package com.cosain.trilo.trip.presentation.trip;
 
-import com.cosain.trilo.common.LoginUser;
+import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.presentation.Login;
+import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.trip.service.TripTitleUpdateService;
 import com.cosain.trilo.trip.application.trip.dto.TripTitleUpdateCommand;
 import com.cosain.trilo.trip.application.trip.dto.factory.TripTitleUpdateCommandFactory;
@@ -22,8 +24,9 @@ public class TripTitleUpdateController {
 
     @PutMapping("/api/trips/{tripId}/title")
     @ResponseStatus(HttpStatus.OK)
-    public TripTitleUpdateResponse updateTrip(@LoginUser User user, @PathVariable Long tripId, @RequestBody TripTitleUpdateRequest request) {
-        Long tripperId = user.getId();
+    @Login
+    public TripTitleUpdateResponse updateTrip(@LoginUser UserPayload userPayload, @PathVariable Long tripId, @RequestBody TripTitleUpdateRequest request) {
+        Long tripperId = userPayload.getId();
 
         TripTitleUpdateCommand updateCommand = tripTitleUpdateCommandFactory.createCommand(request.getTitle());
         tripTitleUpdateService.updateTripTitle(tripId, tripperId, updateCommand);
